+++ docs/CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog
 - Enh #82: Simply Tt menu items
 - Fix #97: [Bug] Images duplicate on drag and drop
 - Enh #43: Make menu accessible
+- Fix #51: Link creation does not copy URL from selected text
 
 1.1.7 (August 31, 2021)
 -----------------------